+++ build_drmtool.sh
@@ -4,18 +4,30 @@
 # This is built and installed drm_tool twice as we need to checkout an older version
 # for panel color profile setting.  Yes it's weird and shouldn't be needed
 # Perhaps i'll clean it up later.  Maybe not. I don't know
-call_chroot "cd /home/ark &&
-  git clone --recursive https://github.com/christianhaitian/drm_tool.git &&
-  cd drm_tool &&
-  make &&
-  strip drm_tool &&
-  cp drm_tool /usr/local/bin/ &&
-  chmod 777 /usr/local/bin/drm_tool &&
-  make clean &&
-  git checkout 1cb5b10b7d529105e33f27388519671ee7ce46f3 &&
-  make &&
-  strip drm_tool &&
-  cp drm_tool /usr/local/bin/panel_drm_tool &&
-  chmod 777 /usr/local/bin/panel_drm_tool
-  "
-sudo rm -rf Arkbuild/home/ark/drm_tool
+if [ -f "Arkbuild_package_cache/${CHIPSET}/drmtool.tar.gz" ] && [ "$(cat Arkbuild_package_cache/${CHIPSET}/drmtool.commit)" == "$(curl -s https://api.github.com/repos/christianhaitian/drm_tool/commits/master | jq -r '.sha')" ]; then
+    sudo tar -xvzpf Arkbuild_package_cache/${CHIPSET}/drmtool.tar.gz
+else
+	call_chroot "cd /home/ark &&
+	  git clone --recursive https://github.com/christianhaitian/drm_tool.git &&
+	  cd drm_tool &&
+	  make &&
+	  strip drm_tool &&
+	  cp drm_tool /usr/local/bin/ &&
+	  chmod 777 /usr/local/bin/drm_tool &&
+	  make clean &&
+	  git checkout 1cb5b10b7d529105e33f27388519671ee7ce46f3 &&
+	  make &&
+	  strip drm_tool &&
+	  cp drm_tool /usr/local/bin/panel_drm_tool &&
+	  chmod 777 /usr/local/bin/panel_drm_tool
+	  "
+	sudo rm -rf Arkbuild/home/ark/drm_tool
+	if [ -f "Arkbuild_package_cache/${CHIPSET}/drmtool.tar.gz" ]; then
+	  sudo rm -f Arkbuild_package_cache/${CHIPSET}/drmtool.tar.gz
+	fi
+	if [ -f "Arkbuild_package_cache/${CHIPSET}/drmtool.commit" ]; then
+	  sudo rm -f Arkbuild_package_cache/${CHIPSET}/drmtool.commit
+	fi
+	sudo tar -czpf Arkbuild_package_cache/${CHIPSET}/drmtool.tar.gz Arkbuild/usr/local/bin/drm_tool Arkbuild/usr/local/bin/panel_drm_tool
+	curl -s https://api.github.com/repos/christianhaitian/drm_tool/commits/master | jq -r '.sha' > Arkbuild_package_cache/${CHIPSET}/drmtool.commit
+fi
